requirements.txt
Flask==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
bcrypt==4.3.0
gunicorn==23.0.0
psycopg[binary]==3.2.13
python-dotenv==1.1.1
